fix: Разбить приложение на модули
* Разбил на модули, т.к. возникли проблемы при внедрении в качестве зависимости
