+++ cars/cars-api/src/main/java/ru/gilko/carsapi/constants/enums/CarType.java
@@ -1,0 +1,8 @@
+package ru.gilko.carsapi.constants.enums;
+
+public enum CarType {
+    SEDAN,
+    SUV,
+    MINIVAN,
+    ROADSTER
+}
