+++ cars/cars-impl/src/main/java/ru/gilko/carsimpl/CarsApplication.java
@@ -1,0 +1,13 @@
+package ru.gilko.carsimpl;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CarsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CarsApplication.class, args);
+	}
+
+}
