+++ cars/cars-impl/src/main/java/ru/gilko/carsimpl/mapper/ModelMapperConfig.java
@@ -1,0 +1,26 @@
+package ru.gilko.carsimpl.mapper;
+
+import org.modelmapper.ModelMapper;
+import org.modelmapper.convention.MatchingStrategies;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ModelMapperConfig {
+    @Bean
+    public ModelMapper modelMapper() {
+        ModelMapper mapper = new ModelMapper();
+
+        setUpMapper(mapper);
+
+        return mapper;
+    }
+
+    private void setUpMapper(ModelMapper mapper) {
+        mapper.getConfiguration()
+                .setMatchingStrategy(MatchingStrategies.STRICT)
+                .setFieldMatchingEnabled(true)
+                .setSkipNullEnabled(true)
+                .setFieldAccessLevel(org.modelmapper.config.Configuration.AccessLevel.PRIVATE);
+    }
+}
