--- cars/src/main/java/ru/gilko/cars/CarsApplication.java
@@ -1,13 +1,0 @@
-package ru.gilko.cars;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CarsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CarsApplication.class, args);
-	}
-
-}
