--- cars/src/main/java/ru/gilko/cars/constants/ControllerUrl.java
@@ -1,8 +1,0 @@
-package ru.gilko.cars.constants;
-
-public class ControllerUrl {
-    private ControllerUrl(){}
-
-    public static final String CARS_URL = "/cars";
-    public static final String CONCRETE_CAR_URL = CARS_URL + "/{carId}";
-}
