--- cars/src/main/java/ru/gilko/cars/constants/enums/CarType.java
@@ -1,8 +1,0 @@
-package ru.gilko.cars.constants.enums;
-
-public enum CarType {
-    SEDAN,
-    SUV,
-    MINIVAN,
-    ROADSTER
-}
