--- cars/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-spring.liquibase.change-log=classpath:db/changelog/db.changelog-master.xml
-spring.liquibase.enabled=true
-spring.datasource.driver-class-name=org.postgresql.Driver
-
-spring.datasource.url=jdbc:postgresql://localhost:5432/cars
-spring.datasource.username=postgres
-spring.datasource.password=postgres
-
-spring.application.name=cars
-server.port=8070
-server.servlet.context-path=/api/v1
